pyproject.toml
[project]
name = "ragalyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Joshua Gao", email = "joshuakgao@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "detect-secrets>=1.5.0",
    "dspy>=3.0.1",
    "faiss-cpu>=1.12.0",
    "hydra-core>=1.3.2",
    "inquirer>=3.4.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-google-genai>=2.1.9",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.30",
    "lightrag-hku>=1.4.6",
    "matplotlib>=3.10.5",
    "omegaconf>=2.3.0",
    "openai<1.100.0",
    "pre-commit>=4.3.0",
    "pypdf>=6.0.0",
    "pytest>=8.4.2",
    "ragas>=0.3.2",
    "ruff>=0.12.12",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.55.2",
]

[project.scripts]
ragalyst = "ragalyst.cli:main"

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"